+++ ads/aqua/base.py
@@ -4,13 +4,26 @@
 # Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/
 
 
+from typing import Dict, List, Union
 import oci
 
 from ads import set_auth
+from ads.aqua.exception import AquaValueError
+from ads.aqua.utils import UNKNOWN, is_valid_ocid, logger
 from ads.common import oci_client as oc
 from ads.common.auth import default_signer
 from ads.common.utils import extract_region
 from ads.config import OCI_ODSC_SERVICE_ENDPOINT, OCI_RESOURCE_PRINCIPAL_VERSION
+from ads.model.datascience_model import DataScienceModel
+from ads.model.deployment.model_deployment import ModelDeployment
+from ads.model.model_metadata import ModelCustomMetadata, ModelProvenanceMetadata, ModelTaxonomyMetadata
+from ads.model.model_version_set import ModelVersionSet
+
+from oci.data_science.models import (
+    Metadata,
+    UpdateModelDetails,
+    UpdateModelProvenanceDetails,
+)
 
 
 class AquaApp:
@@ -49,3 +62,103 @@ class AquaApp:
             list_func_ref,
             **kwargs,
         ).data
+    
+    def update_model(
+        self,
+        model_id: str,
+        update_model_details: UpdateModelDetails
+    ):
+        self.ds_client.update_model(
+            model_id=model_id,
+            update_model_details=update_model_details
+        )
+
+    def update_model_provenance(
+        self,
+        model_id: str,
+        update_model_provenance_details: UpdateModelProvenanceDetails
+    ):
+        self.ds_client.update_model_provenance(
+            model_id=model_id,
+            update_model_provenance_details=update_model_provenance_details
+        )
+    
+    @staticmethod
+    def get_source(source_id: str) -> Union[ModelDeployment, DataScienceModel]:
+        if is_valid_ocid(source_id):
+            if "datasciencemodeldeployment" in source_id:
+                return ModelDeployment.from_id(source_id)
+            elif "datasciencemodel" in source_id:
+                return DataScienceModel.from_id(source_id)
+        
+        raise AquaValueError(
+            f"Invalid source {source_id}. "
+            "Specify either a model or model deployment id."
+        )
+    
+    @staticmethod
+    def create_model_version_set(
+        model_version_set_id: str=None,
+        model_version_set_name: str=None,
+        description: str=None,
+        compartment_id: str=None,
+        project_id: str=None,
+        **kwargs
+    ) -> tuple:
+        if not model_version_set_id:
+            try:
+                model_version_set = ModelVersionSet.from_name(
+                    name=model_version_set_name,
+                    compartment_id=compartment_id,
+                )
+            except:
+                logger.debug(
+                    f"Model version set {model_version_set_name} doesn't exist. "
+                    "Creating new model version set."
+                )
+                model_version_set = (
+                    ModelVersionSet()
+                    .with_compartment_id(compartment_id)
+                    .with_project_id(project_id)
+                    .with_name(model_version_set_name)
+                    .with_description(description)
+                    # TODO: decide what parameters will be needed
+                    .create(**kwargs)
+                )
+                logger.debug(
+                    f"Successfully created model version set {model_version_set_name} with id {model_version_set.id}."
+                )
+            return (model_version_set.id, model_version_set_name)
+        else:
+            model_version_set = ModelVersionSet.from_id(model_version_set_id)
+            return (model_version_set_id, model_version_set.name)
+    
+    @staticmethod
+    def create_model_catalog(
+        display_name: str,
+        description: str,
+        model_version_set_id: str,
+        model_custom_metadata: Union[ModelCustomMetadata, Dict],
+        model_taxonomy_metadata: Union[ModelTaxonomyMetadata, Dict],
+        compartment_id: str,
+        project_id: str,
+        **kwargs
+    ) -> DataScienceModel:
+        model = (
+            DataScienceModel()
+            .with_compartment_id(compartment_id)
+            .with_project_id(project_id)
+            .with_display_name(display_name)
+            .with_description(description)
+            .with_model_version_set_id(model_version_set_id)
+            .with_custom_metadata_list(model_custom_metadata)
+            .with_defined_metadata_list(model_taxonomy_metadata)
+            .with_provenance_metadata(
+                ModelProvenanceMetadata(training_id=UNKNOWN)
+            )
+            # TODO: decide what parameters will be needed
+            .create(
+                **kwargs,
+            )
+        )
+        return model
